Extend the range of protection against all-zeros substance amount fractions
	- use intermediate cast to bigint rather than int

+++ gnumed/gnumed/server/sql/v16-v17/fixups/v17-ref-consumable_substance-fixup.sql
@@ -28,7 +28,7 @@ BEGIN
 	if _fraction <> 0 then
 		return _numeric_value;
 	end if;
-	return _numeric_value::integer::numeric;
+	return _numeric_value::bigint::numeric;
 END;';
